requirements.txt
openai
python-dotenv
fastapi
uvicorn
google-generativeai
asyncpg
websockets
google-cloud-speech